+++ pkg/executor/server/handlers.go
@@ -6,12 +6,12 @@ import (
 	"net/http"
 
 	"github.com/gofiber/fiber/v2"
-	"github.com/kubeshop/kubtest/internal/pkg/server"
 	"github.com/kubeshop/kubtest/pkg/api/kubtest"
 	"github.com/kubeshop/kubtest/pkg/executor/repository"
+	"github.com/kubeshop/kubtest/pkg/server"
 )
 
-func StartExecution(s *server.HTTPServer, repository repository.Repository) fiber.Handler {
+func StartExecution(s server.HTTPServer, repository repository.Repository) fiber.Handler {
 	return func(c *fiber.Ctx) error {
 
 		var request kubtest.ExecutionRequest
@@ -45,7 +45,7 @@ func StartExecution(s *server.HTTPServer, repository repository.Repository) fibe
 	}
 }
 
-func GetExecution(s *server.HTTPServer, repository repository.Repository) fiber.Handler {
+func GetExecution(s server.HTTPServer, repository repository.Repository) fiber.Handler {
 	return func(c *fiber.Ctx) error {
 		execution, err := repository.Get(context.Background(), c.Params("id"))
 		if err != nil {
